Update pre_process.py
I suggest to shift the origin time to the day of ovulation for estimating the probability of occurrence of a certain symptom for each specific id. The reason is explained in the following considerations:
- a menstrual cycle is made up of three phases: 1) the follicular phase; 2) the day of ovulation; and 3) the luteal phase;
- the luteal phase length is in average 14 days (+/- 1 or 2 days) and it is almost a constant for each woman;
- the variability in cycle length is attributable to the follicular phase length.  
In this way the predictions made for an id and based on an averaged cycle length (made over a number of cycles) are gaining in accuracy.